+++ src/wb_node.ts
@@ -32,6 +32,7 @@ import {
   SortCallback,
   NodeToDictCallback,
   WbNodeData,
+  TristateType,
 } from "./types";
 import {
   iconMap,
@@ -47,31 +48,18 @@ import {
 } from "./common";
 import { Deferred } from "./deferred";
 
-/** Top-level properties that can be passed with `data`. */
+/** WunderbaumNode properties that can be passed with source data.
+ * (Any other source properties will be stored as `node.data.PROP`.)
+ */
 const NODE_PROPS = new Set<string>([
-  // TODO: use NODE_ATTRS instead?
-  "classes",
-  "expanded",
-  "icon",
-  "key",
-  "lazy",
-  "refKey",
-  "selected",
-  "title",
-  "tooltip",
-  "type",
-]);
-
-const NODE_ATTRS = new Set<string>([
   "checkbox",
-  "expanded",
   "classes",
-  "folder",
+  "expanded",
   "icon",
   "iconTooltip",
   "key",
   "lazy",
-  "partsel",
+  "_partsel",
   "radiogroup",
   "refKey",
   "selected",
@@ -80,9 +68,16 @@ const NODE_ATTRS = new Set<string>([
   "tooltip",
   "type",
   "unselectable",
-  "unselectableIgnore",
-  "unselectableStatus",
+  // "unselectableIgnore",
+  // "unselectableStatus",
 ]);
+/** WunderbaumNode properties that will be returned by `node.toDict()`.)
+ */
+const NODE_DICT_PROPS = new Set<string>(NODE_PROPS);
+NODE_DICT_PROPS.delete("_partsel");
+NODE_DICT_PROPS.delete("unselectable");
+// NODE_DICT_PROPS.delete("unselectableIgnore");
+// NODE_DICT_PROPS.delete("unselectableStatus");
 
 /**
  * A single tree node.
@@ -111,6 +106,7 @@ export class WunderbaumNode {
   public readonly refKey: string | undefined = undefined;
   public children: WunderbaumNode[] | null = null;
   public checkbox?: boolean;
+  public radiogroup?: boolean;
   /** If true, (in grid mode) no cells are rendered, except for the node title.*/
   public colspan?: boolean;
   public icon?: boolean | string;
@@ -119,8 +115,11 @@ export class WunderbaumNode {
    * @see {@link isExpandable}, {@link isExpanded}, {@link setExpanded}. */
   public expanded: boolean = false;
   /** Selection state.
-   * @see {@link isSelected}, {@link setSelected}. */
+   * @see {@link isSelected}, {@link setSelected}, {@link toggleSelected}. */
   public selected: boolean = false;
+  public unselectable?: boolean;
+  // public unselectableStatus?: boolean;
+  // public unselectableIgnore?: boolean;
   public type?: string;
   public tooltip?: string;
   /** Additional classes added to `div.wb-row`.
@@ -149,22 +148,32 @@ export class WunderbaumNode {
   constructor(tree: Wunderbaum, parent: WunderbaumNode, data: any) {
     util.assert(!parent || parent.tree === tree);
     util.assert(!data.children);
+
     this.tree = tree;
     this.parent = parent;
+
     this.key = "" + (data.key ?? ++WunderbaumNode.sequence);
     this.title = "" + (data.title ?? "<" + this.key + ">");
-
     data.refKey != null ? (this.refKey = "" + data.refKey) : 0;
-    data.statusNodeType != null
-      ? (this.statusNodeType = "" + data.statusNodeType)
-      : 0;
     data.type != null ? (this.type = "" + data.type) : 0;
-    data.checkbox != null ? (this.checkbox = !!data.checkbox) : 0;
-    data.colspan != null ? (this.colspan = !!data.colspan) : 0;
     this.expanded = data.expanded === true;
     data.icon != null ? (this.icon = data.icon) : 0;
     this.lazy = data.lazy === true;
+    data.statusNodeType != null
+      ? (this.statusNodeType = "" + data.statusNodeType)
+      : 0;
+    data.colspan != null ? (this.colspan = !!data.colspan) : 0;
+
+    // Selection
+    data.checkbox != null ? (this.checkbox = !!data.checkbox) : 0;
+    data.radiogroup != null ? (this.radiogroup = !!data.radiogroup) : 0;
     this.selected = data.selected === true;
+    data.unselectable === true ? (this.unselectable = true) : 0;
+    // data.unselectableStatus != null
+    //   ? (this.unselectableStatus = !!data.unselectableStatus)
+    //   : 0;
+    // data.unselectableIgnore === true ? (this.unselectableIgnore = true) : 0;
+
     if (data.classes) {
       this.setClass(data.classes);
     }
@@ -310,13 +319,16 @@ export class WunderbaumNode {
         // insert nodeList after children[pos]
         this.children.splice(pos, 0, ...nodeList);
       }
-      // TODO:
-      // if (tree.options.selectMode === 3) {
-      //   this.fixSelection3FromEndNodes();
-      // }
       // this.triggerModifyChild("add", nodeList.length === 1 ? nodeList[0] : null);
       tree.update(ChangeType.structure);
     } finally {
+      // if (tree.options.selectMode === "hier") {
+      //   if (this.parent && this.parent.children) {
+      //     this.fixSelection3FromEndNodes();
+      //   } else {
+      //     // my happen when loading __root__;
+      //   }
+      // }
       tree.enableUpdate(true);
     }
     // if(isTopCall && loadLazy){
@@ -421,7 +433,7 @@ export class WunderbaumNode {
     const minExpandLevel = this.tree.options.minExpandLevel;
     let { depth = 99, loadLazy, force } = options ?? {};
 
-    const expand_opts = {
+    const expandOpts = {
       scrollIntoView: false,
       force: force,
       loadLazy: loadLazy,
@@ -445,7 +457,7 @@ export class WunderbaumNode {
             // Node is collapsed and may be expanded (i.e. has children or is lazy)
             // Expanding may be async, so we store the promise.
             // Also the recursion is delayed until expansion finished.
-            const p = cn.setExpanded(true, expand_opts);
+            const p = cn.setExpanded(true, expandOpts);
             promises.push(p);
             p.then(async () => {
               await _iter(cn, level_1);
@@ -459,7 +471,7 @@ export class WunderbaumNode {
           // Collapsing is always synchronous, so no promises required
           if (!minExpandLevel || force || cn.getLevel() > minExpandLevel) {
             // Do not collapse until minExpandLevel
-            cn.setExpanded(false, expand_opts);
+            cn.setExpanded(false, expandOpts);
           }
           _iter(cn, level_1); // recursion, even if cn was already collapsed
         }
@@ -885,9 +897,11 @@ export class WunderbaumNode {
     return this.tree.root === this;
   }
 
-  /** Return true if this node is selected, i.e. the checkbox is set. */
-  isSelected(): boolean {
-    return !!this.selected;
+  /** Return true if this node is selected, i.e. the checkbox is set.
+   * `undefined` if partly selected (tri-state), false otherwise.
+   */
+  isSelected(): TristateType {
+    return this.selected ? true : this._partsel ? undefined : false;
   }
 
   /** Return true if this node is a temporarily generated system node like
@@ -990,6 +1004,9 @@ export class WunderbaumNode {
         tree.data[key] = value;
         tree.logDebug(`Add source.${key} to tree.data.${key}`);
       }
+    }
+    if (tree.options.selectMode === "hier") {
+      this.fixSelection3FromEndNodes();
     }
 
     this._callEvent("load");
@@ -1317,7 +1334,7 @@ export class WunderbaumNode {
       // Fix node.tree for all source nodes
       // 	util.assert(false, "Cross-tree move is not yet implemented.");
       this.logWarn("Cross-tree moveTo is experimental!");
-      this.visit(function (n) {
+      this.visit((n) => {
         // TODO: fix selection state and activation, ...
         n.tree = targetNode.tree;
       }, true);
@@ -1514,7 +1531,8 @@ export class WunderbaumNode {
   protected _render_markup(opts: RenderOptions) {
     const tree = this.tree;
     const treeOptions = tree.options;
-    const checkbox = this.getOption("checkbox") !== false;
+    const checkbox = this.getOption("checkbox");
+    // const checkbox = this.getOption("checkbox") !== false;
     const columns = tree.columns;
     const level = this.getLevel();
     let elem: HTMLElement;
@@ -1553,6 +1571,9 @@ export class WunderbaumNode {
     if (checkbox) {
       checkboxSpan = document.createElement("i");
       checkboxSpan.classList.add("wb-checkbox");
+      if (checkbox === "radio" || this.parent.radiogroup) {
+        checkboxSpan.classList.add("wb-radio");
+      }
       nodeElem.appendChild(checkboxSpan);
       ofsTitlePx += ICON_WIDTH;
     }
@@ -1739,6 +1760,7 @@ export class WunderbaumNode {
     this.expanded ? rowClasses.push("wb-expanded") : 0;
     this.lazy ? rowClasses.push("wb-lazy") : 0;
     this.selected ? rowClasses.push("wb-selected") : 0;
+    this._partsel ? rowClasses.push("wb-partsel") : 0;
     this === tree.activeNode ? rowClasses.push("wb-active") : 0;
     this === tree.focusNode ? rowClasses.push("wb-focus") : 0;
     this._errorInfo ? rowClasses.push("wb-error") : 0;
@@ -1779,11 +1801,26 @@ export class WunderbaumNode {
       }
     }
     if (checkboxSpan) {
-      if (this.selected) {
-        checkboxSpan.className = "wb-checkbox " + iconMap.checkChecked;
+      let cbclass = "wb-checkbox ";
+      if (this.parent.radiogroup) {
+        cbclass += "wb-radio ";
+        if (this.selected) {
+          cbclass += iconMap.radioChecked;
+          // } else if (this._partsel) {
+          //   cbclass += iconMap.radioUnknown;
+        } else {
+          cbclass += iconMap.radioUnchecked;
+        }
       } else {
-        checkboxSpan.className = "wb-checkbox " + iconMap.checkUnchecked;
+        if (this.selected) {
+          cbclass += iconMap.checkChecked;
+        } else if (this._partsel) {
+          cbclass += iconMap.checkUnknown;
+        } else {
+          cbclass += iconMap.checkUnchecked;
+        }
       }
+      checkboxSpan.className = cbclass;
     }
     // Fix active cell in cell-nav mode
     if (!opts.isNew) {
@@ -1874,7 +1911,7 @@ export class WunderbaumNode {
   toDict(recursive = false, callback?: NodeToDictCallback): WbNodeData {
     const dict: any = {};
 
-    NODE_ATTRS.forEach((propName: string) => {
+    NODE_DICT_PROPS.forEach((propName: string) => {
       const val = (<any>this)[propName];
 
       if (val instanceof Set) {
@@ -2120,14 +2157,235 @@ export class WunderbaumNode {
     this.tree.update(change, this);
   }
 
+  /**
+   * Return an array of selected nodes.
+   * @param stopOnParents only return the topmost selected node (useful with selectMode 'hier')
+   */
+  getSelectedNodes(stopOnParents: boolean = false): WunderbaumNode[] {
+    let nodeList: WunderbaumNode[] = [];
+    this.visit((node) => {
+      if (node.selected) {
+        nodeList.push(node);
+        if (stopOnParents === true) {
+          return "skip"; // stop processing this branch
+        }
+      }
+    });
+    return nodeList;
+  }
+
+  /** Toggle the check/uncheck state. */
+  toggleSelected(options?: SetSelectedOptions): TristateType {
+    let flag = this.isSelected();
+    if (flag === undefined) {
+      flag = this._anySelectable();
+    } else {
+      flag = !flag;
+    }
+    return this.setSelected(flag, options);
+  }
+
+  /** Return true if at least on selectable descendant end-node is unselected. @internal */
+  _anySelectable(): boolean {
+    let found = false;
+    this.visit((node) => {
+      if (
+        node.selected === false &&
+        !node.unselectable &&
+        !node.hasChildren() &&
+        !node.parent.radiogroup
+      ) {
+        found = true;
+        return false; // Stop iteration
+      }
+    });
+    return found;
+  }
+
+  /* Apply selection state to a single node. */
+  protected _changeSelectStatusProps(state: TristateType): boolean {
+    let changed = false;
+    switch (state) {
+      case false:
+        changed = this.selected || this._partsel;
+        this.selected = false;
+        this._partsel = false;
+        break;
+      case true:
+        changed = !this.selected || !this._partsel;
+        this.selected = true;
+        this._partsel = true;
+        break;
+      case undefined:
+        changed = this.selected || !this._partsel;
+        this.selected = false;
+        this._partsel = true;
+        break;
+      default:
+        util.error(`Invalid state: ${state}`);
+    }
+    if (changed) {
+      this.update();
+    }
+    return changed;
+  }
+  /**
+   * Fix selection status, after this node was (de)selected in `selectMode: 'hier'`.
+   * This includes (de)selecting all descendants.
+   */
+  fixSelection3AfterClick(opts?: SetSelectedOptions): void {
+    const force = !!opts?.force;
+    let flag = this.isSelected();
+
+    this.visit((node) => {
+      if (node.radiogroup) {
+        return "skip"; // Don't (de)select this branch
+      }
+      if (force || !node.getOption("unselectable")) {
+        node._changeSelectStatusProps(flag);
+      }
+    });
+    this.fixSelection3FromEndNodes();
+  }
+
+  /**
+   * Fix selection status for multi-hier mode.
+   * Only end-nodes are considered to update the descendants branch and parents.
+   * Should be called after this node has loaded new children or after
+   * children have been modified using the API.
+   */
+  fixSelection3FromEndNodes(opts?: SetSelectedOptions): void {
+    const force = !!opts?.force;
+    util.assert(
+      this.tree.options.selectMode === "hier",
+      "expected selectMode 'hier'"
+    );
+
+    // Visit all end nodes and adjust their parent's `selected` and `_partsel`
+    // attributes. Return selection state true, false, or undefined.
+    const _walk = (node: WunderbaumNode) => {
+      let state;
+      const children = node.children;
+
+      if (children && children.length) {
+        // check all children recursively
+        let allSelected = true;
+        let someSelected = false;
+
+        for (let i = 0, l = children.length; i < l; i++) {
+          const child = children[i];
+          // the selection state of a node is not relevant; we need the end-nodes
+          const s = _walk(child);
+          if (s !== false) {
+            someSelected = true;
+          }
+          if (s !== true) {
+            allSelected = false;
+          }
+        }
+        state = allSelected ? true : someSelected ? undefined : false;
+      } else {
+        // This is an end-node: simply report the status
+        state = !!node.selected;
+      }
+      // #939: Keep a `_partsel` flag that was explicitly set on a lazy node
+      if (
+        node._partsel &&
+        !node.selected &&
+        node.lazy &&
+        node.children == null
+      ) {
+        state = undefined;
+      }
+      if (force || !node.getOption("unselectable")) {
+        node._changeSelectStatusProps(state);
+      }
+      return state;
+    };
+    _walk(this);
+
+    // Update parent's state
+    this.visitParents((node) => {
+      let state;
+      const children = node.children!;
+      let allSelected = true;
+      let someSelected = false;
+
+      for (let i = 0, l = children.length; i < l; i++) {
+        const child = children[i];
+
+        state = !!child.selected;
+        // When fixing the parents, we trust the sibling status (i.e. we don't recurse)
+        if (state || child._partsel) {
+          someSelected = true;
+        }
+        if (!state) {
+          allSelected = false;
+        }
+      }
+      state = allSelected ? true : someSelected ? undefined : false;
+      node._changeSelectStatusProps(state);
+    });
+  }
+
   /** Modify the check/uncheck state. */
-  setSelected(flag: boolean = true, options?: SetSelectedOptions) {
-    const prev = this.selected;
-    if (!!flag !== prev) {
+  setSelected(
+    flag: boolean = true,
+    options?: SetSelectedOptions
+  ): TristateType {
+    const tree = this.tree;
+    const sendEvents = !options?.noEvents; // Default: send events
+    const prev = this.isSelected();
+    const isRadio = this.parent && this.parent.radiogroup;
+    const selectMode = tree.options.selectMode;
+    const canSelect = options?.force || !this.getOption("unselectable");
+
+    flag = !!flag;
+    // this.logDebug(`setSelected(${flag})`, this);
+    if (!canSelect) {
+      return prev;
+    }
+    if (options?.propagateDown && selectMode === "multi") {
+      tree.runWithDeferredUpdate(() => {
+        this.visit((node) => {
+          node.setSelected(flag);
+        });
+      });
+      return prev;
+    }
+
+    if (
+      flag === prev ||
+      (sendEvents && this._callEvent("beforeSelect", { flag: flag }) === false)
+    ) {
+      return prev;
+    }
+
+    tree.runWithDeferredUpdate(() => {
+      if (isRadio) {
+        // Radiobutton Group
+        if (!flag && !options?.force) {
+          return prev; // don't uncheck radio buttons
+        }
+        for (let sibling of this.parent.children!) {
+          sibling.selected = sibling === this;
+        }
+      } else {
+        this.selected = flag;
+        if (selectMode === "hier") {
+          this.fixSelection3AfterClick();
+        } else if (selectMode === "single") {
+          tree.visit((n) => {
+            n.selected = false;
+          });
+        }
+      }
+    });
+
+    if (sendEvents) {
       this._callEvent("select", { flag: flag });
     }
-    this.selected = !!flag;
-    this.update();
+    return prev;
   }
 
   /** Display node status (ok, loading, error, noData) using styles and a dummy child node. */
@@ -2296,7 +2554,7 @@ export class WunderbaumNode {
   }
 
   /**
-   * Call `callback(node)` for all child nodes in hierarchical order (depth-first, pre-order).
+   * Call `callback(node)` for all descendant nodes in hierarchical order (depth-first, pre-order).
    *
    * Stop iteration, if fn() returns false. Skip current branch, if fn()
    * returns "skip".<br>